refactor: move ssoData to package
